Consolidate window size fixtures and correct the measurement KDoc
One shared fixture in PreviewHelpers drives both the Compose previews
(moved to LoadedPaywallComponentsPreviews per convention) and the Paparazzi
test, now one config per outcome instead of six. WindowWidthRule's KDoc now
states the Android behavior: the app window, with sheets and dialogs seeing
their enclosing window's size.
