+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/components/LoadedPaywallComponentsPreviews.kt
@@ -1429,3 +1429,15 @@ private fun LoadedPaywallComponents_Preview_LargeFooter() {
         modifier = Modifier.fillMaxSize(),
     )
 }
+
+@Preview(name = "WindowSize - phone stacked", device = "spec:width=411dp,height=891dp", showSystemUi = true)
+@Preview(name = "WindowSize - landscape stacked (height floor)", device = "spec:width=891dp,height=411dp", showSystemUi = true)
+@Preview(name = "WindowSize - tablet side by side", device = "spec:width=1280dp,height=800dp,dpi=240", showSystemUi = true)
+@Composable
+private fun LoadedPaywallComponents_Preview_WindowSizeConditions() {
+    LoadedPaywallComponents(
+        state = previewWindowSizeConditionsState(),
+        clickHandler = { },
+        modifier = Modifier.fillMaxSize(),
+    )
+}
